Refactor ntpd to support non-root execution
To support running ntpd as a non-root user,
split the original ntpd.c into two files:
ntpd.c and ntpd_nonroot.c. This results in
two separate executables: ntpd and ntpd_nonroot.
The ntpd binary is responsible for setting
process priority and resource limits. After
performing these privileged operations, it
drops root privilege and then uses execvp to
launch ntpd_nonroot. This design allows users
to run ntpd_nonroot directly as a non-root
user to utilize NTP functionality.
